# bigint: small to_str_radix optimization