@cuviper 2016-01-05
Huh.  I know this table lookup proved faster than the old `to_digit`, which used a match on ranges.  If your single branch is faster still, so be it!  Especially for radix < 10, I'll bet that branch predicts true really well.

CI is failing rust nightly in building rand. :/  rust-lang-nursery/rand#91 rust-lang/rust#30713