@cuviper 2016-01-12
Nightly vs. rand should be resolved now -- let's give it a whirl...

@homu r+ d2b23d0