@homu 2016-01-13
:sunny: Test successful - [status](https://travis-ci.org/rust-num/num/builds/101983410)